+++ CoreDisc/Views/Onboarding/LoginView.swift
@@ -6,9 +6,6 @@
 //
 
 import SwiftUI
-import KakaoSDKCommon
-import KakaoSDKAuth
-import KakaoSDKUser
 import Moya
 
 struct LoginView: View {
@@ -138,77 +135,17 @@ struct LoginView: View {
             
             HStack(spacing: 43){
                 Image(.imgNaver)
-                KakaoLoginButton(isLoginSuccess: $isLoginSuccess)
+                
+                Button(action: {
+                    viewModel.kakaoLogin()
+                }) {
+                    Image(.imgKakao)
+                }
+                
                 Image(.imgGoogle)
             }
         }
         .padding(.horizontal, 41)
-    }
-}
-
-// 카카오 로그인 버튼
-struct KakaoLoginButton: View {
-    @Binding var isLoginSuccess: Bool
-    
-    var body: some View {
-        Button(action: {
-            let provider = APIManager.shared.createProvider(for: AuthRouter.self)
-            func kakaoLogin(accessToken: String) {
-                provider.request(.postKakao(accessToken: accessToken)) { result in
-                    switch result {
-                    case .success(let response):
-                        do {
-                            let decodedData = try JSONDecoder().decode(KakaoResponse.self, from: response.data)
-                            
-                            let userInfo = UserInfo(
-                                accessToken: decodedData.result.accessToken,
-                                refreshToken: decodedData.result.refreshToken
-                            )
-                            
-                            // 토큰 KeyChain에 저장
-                            let saved = KeychainManager.standard.saveSession(userInfo, for: "appNameUser")
-                            if saved {
-                                print("Token 저장 성공: \(String(describing: userInfo.accessToken))")
-                                isLoginSuccess = true // 화면 전환
-                            } else {
-                                print("Token 저장 실패")
-                            }
-                        } catch {
-                            print("KakaoResponse 디코더 오류: \(error)")
-                        }
-                    case .failure(let error):
-                        print("Kakao Login Error: \(error)")
-                    }
-                }
-            }
-            
-            // 카카오톡 실행 가능 여부 확인
-            if UserApi.isKakaoTalkLoginAvailable() {
-                // 카카오톡 로그인
-                UserApi.shared.loginWithKakaoTalk { oauthToken, error in
-                    if let error = error {
-                        print(error)
-                    } else {
-                        if let accessToken = oauthToken?.accessToken {
-                            kakaoLogin(accessToken: accessToken)
-                        }
-                    }
-                }
-            } else {
-                // 카카오계정 로그인
-                UserApi.shared.loginWithKakaoAccount { oauthToken, error in
-                    if let error = error {
-                        print(error)
-                    } else {
-                        if let accessToken = oauthToken?.accessToken {
-                            kakaoLogin(accessToken: accessToken)
-                        }
-                    }
-                }
-            }
-        }) {
-            Image(.imgKakao)
-        }
     }
 }
